refactor(terminal): split panel features
